fix: fix bug when set is not over at 7-5 games. Also tweak parsePageNumber in DataValidationUtil

+++ src/main/java/com/dimasukimas/tennisscoreboard/util/DataValidationUtil.java
@@ -40,7 +40,7 @@ public class DataValidationUtil {
     public static int parsePageNumber(String id) {
         int pageNumber;
         try {
-            pageNumber = Integer.parseInt(id);
+            pageNumber = (int) Long.parseLong(id);
         } catch (IllegalArgumentException e) {
             log.warn("Failed to parse page number {}", id, e);
             throw new IllegalArgumentException("Incorrect page number", e);
